Unify TimeStepper interface (#122)
* Uses array-of-constant-or-zero more programmatically for Butcher tableaux & Galerkin
* Time derivatives push through list tensor, simplifying substitutions.
* simplify substitutions
* Provide some base classes to enable code reuse 
* Update file structure
* refactor bounds constraints without data movement, just build the stages by referencing dat objects
