+++ irksome/__init__.py
@@ -1,3 +1,4 @@
+from .bcs import BoundsConstrainedDirichletBC  # noqa: F401
 from .ButcherTableaux import Alexander      # noqa: F401
 from .ButcherTableaux import BackwardEuler  # noqa: F401
 from .ButcherTableaux import GaussLegendre  # noqa: F401
@@ -13,12 +14,12 @@ from .ars_dirk_imex_tableaux import ARS_DIRK_IMEX  # noqa: F401
 from .sspk_tableau import SSPK_DIRK_IMEX  # noqa: F401
 from .sspk_tableau import SSPButcherTableau  # noqa: F401
 from .dirk_stepper import DIRKTimeStepper   # noqa: F401
-from .getForm import getForm                # noqa: F401
+from .stage_derivative import getForm       # noqa: F401
 from .imex import RadauIIAIMEXMethod        # noqa: F401
 from .imex import DIRKIMEXMethod            # noqa: F401
 from .pc import RanaBase, RanaDU, RanaLD    # noqa: F401
 from .pc import IRKAuxiliaryOperatorPC      # noqa: F401
-from .stage import StageValueTimeStepper    # noqa: F401
+from .stage_value import StageValueTimeStepper  # noqa: F401
 from .stepper import TimeStepper            # noqa: F401
 from .tools import MeshConstant             # noqa: F401
 from .wso_dirk_tableaux import WSODIRK      # noqa: F401
