+++ irksome/dirk_stepper.py
@@ -5,7 +5,7 @@ from firedrake import NonlinearVariationalSolver as NLVS
 from ufl.constantvalue import as_ufl
 
 from .deriv import TimeDerivative
-from .tools import component_replace, replace, MeshConstant
+from .tools import component_replace, replace, MeshConstant, vecconst
 from .bcs import bc2space
 
 
@@ -30,8 +30,8 @@ def getFormDIRK(F, ks, butch, t, dt, u0, bcs=None):
     c = MC.Constant(1.0)
     a = MC.Constant(1.0)
 
-    repl = {t: t+c*dt,
-            u0: g + dt * a * k,
+    repl = {t: t + c * dt,
+            u0: g + k * (a * dt),
             TimeDerivative(u0): k}
     stage_F = component_replace(F, repl)
 
@@ -45,20 +45,16 @@ def getFormDIRK(F, ks, butch, t, dt, u0, bcs=None):
                          dtype=object)
     d_val = MC.Constant(1.0)
     for bc in bcs:
-        bcarg = as_ufl(bc._original_arg)
-        bcarg_stage = replace(bcarg, {t: t+c*dt})
-        if bcarg_stage == 0:
+        bcarg = bc._original_arg
+        if bcarg == 0:
             # Homogeneous BC, just zero out stage dofs
-            bcnew.append(bc.reconstruct(g=0))
-            continue
-
-        gdat = bcarg_stage - bc2space(bc, u0)
-        for i in range(num_stages):
-            gdat -= dt*a_vals[i]*bc2space(bc, ks[i])
-
-        gdat /= dt*d_val
-
-        bcnew.append(bc.reconstruct(g=gdat))
+            bcnew.append(bc)
+        else:
+            bcarg_stage = replace(as_ufl(bcarg), {t: t+c*dt})
+            gdat = bcarg_stage - bc2space(bc, u0)
+            gdat -= sum(bc2space(bc, ks[i]) * (a_vals[i] * dt) for i in range(num_stages))
+            gdat /= d_val * dt
+            bcnew.append(bc.reconstruct(g=gdat))
 
     return stage_F, (k, g, a, c), bcnew, (a_vals, d_val)
 
@@ -98,6 +94,8 @@ class DIRKTimeStepper:
         if butcher_tableau.is_explicit:
             self.AAb = self.AAb[1:]
             self.CCone = self.CCone[1:]
+        self.AA = vecconst(butcher_tableau.A)
+        self.BB = vecconst(butcher_tableau.b)
 
         self.V = V = u0.function_space()
         self.u0 = u0
@@ -154,23 +152,15 @@ class DIRKTimeStepper:
         k, g, a, c = self.kgac
         ks = self.ks
         u0 = self.u0
-        dtc = float(self.dt)
-        bt = self.butcher_tableau
-        AA = bt.A
-        BB = bt.b
-        gsplit = g.subfunctions
+        dt = self.dt
         for i in range(self.num_stages):
             # compute the already-known part of the state in the
             # variational form
-            g.assign(u0)
-            for j in range(i):
-                ksplit = ks[j].subfunctions
-                for (gbit, kbit) in zip(gsplit, ksplit):
-                    gbit += dtc * float(AA[i, j]) * kbit
+            g.assign(sum((ks[j] * (self.AA[i, j] * dt) for j in range(i)), u0))
 
             # update BC constants for the variational problem
             self.update_bc_constants(i, c)
-            a.assign(AA[i, i])
+            a.assign(self.AA[i, i])
 
             # solve new variational problem, stash the computed
             # stage value.
@@ -181,15 +171,12 @@ class DIRKTimeStepper:
             # former is probably optimal, we hope for the best with
             # the latter.
             self.solver.solve()
-            mysnes = self.solver.snes
-            self.num_nonlinear_iterations += mysnes.getIterationNumber()
-            self.num_linear_iterations += mysnes.getLinearSolveIterations()
+            self.num_nonlinear_iterations += self.solver.snes.getIterationNumber()
+            self.num_linear_iterations += self.solver.snes.getLinearSolveIterations()
             ks[i].assign(k)
 
         # update the solution with now-computed stage values.
-        for i in range(self.num_stages):
-            for (u0bit, kbit) in zip(u0.subfunctions, ks[i].subfunctions):
-                u0bit += dtc * float(BB[i]) * kbit
+        u0 += sum(ks[i] * (self.BB[i] * dt) for i in range(self.num_stages))
 
         self.num_steps += 1
