+++ irksome/discontinuous_galerkin_stepper.py
@@ -1,20 +1,17 @@
-from functools import reduce
 from FIAT import (Bernstein, DiscontinuousElement,
                   DiscontinuousLagrange,
                   IntegratedLegendre, Lagrange,
                   make_quadrature, ufc_simplex)
-from operator import mul
-from ufl import zero
 from ufl.constantvalue import as_ufl
+from .base_time_stepper import StageCoupledTimeStepper
 from .bcs import stage2spaces4bc
 from .manipulation import extract_terms, strip_dt_form
-from .tools import getNullspace, component_replace, replace
+from .tools import component_replace, replace, vecconst
 import numpy as np
-from firedrake import as_vector, Constant, dot, TestFunction, Function, NonlinearVariationalProblem as NLVP, NonlinearVariationalSolver as NLVS
-from firedrake.dmhooks import pop_parent, push_parent
+from firedrake import TestFunction
 
 
-def getFormDiscGalerkin(F, L, Q, t, dt, u0, bcs=None, nullspace=None):
+def getFormDiscGalerkin(F, L, Q, t, dt, u0, stages, bcs=None):
 
     """Given a time-dependent variational form, trial and test spaces, and
     a quadrature rule, produce UFL for the Discontinuous Galerkin-in-Time method.
@@ -29,6 +26,7 @@ def getFormDiscGalerkin(F, L, Q, t, dt, u0, bcs=None, nullspace=None):
          The user may adjust this value between time steps.
     :arg u0: a :class:`Function` referring to the state of
          the PDE system at time `t`
+    :arg stages: a :class:`Function` representing the stages to be solved for.
     :arg bcs: optionally, a :class:`DirichletBC` object (or iterable thereof)
          containing (possibly time-dependent) boundary conditions imposed
          on the system.
@@ -41,11 +39,8 @@ def getFormDiscGalerkin(F, L, Q, t, dt, u0, bcs=None, nullspace=None):
     On output, we return a tuple consisting of four parts:
 
        - Fnew, the :class:`Form` corresponding to the DG-in-Time discretized problem
-       - UU, the :class:`Function` representing the stages to be solved for
        - `bcnew`, a list of :class:`firedrake.DirichletBC` objects to be posed
          on the Galerkin-in-time solution,
-       - 'nspnew', the :class:`firedrake.MixedVectorSpaceBasis` object
-         that represents the nullspace of the coupled system
     """
     assert Q.ref_el.get_spatial_dimension() == 1
     assert L.get_reference_element() == Q.ref_el
@@ -54,15 +49,9 @@ def getFormDiscGalerkin(F, L, Q, t, dt, u0, bcs=None, nullspace=None):
     V = v.function_space()
     assert V == u0.function_space()
 
-    vecconst = Constant
-
     num_stages = L.space_dimension()
-
-    Vbig = reduce(mul, (V for _ in range(num_stages)))
-
-    VV = TestFunction(Vbig)
-    UU = Function(Vbig)
-
+    Vbig = stages.function_space()
+    test = TestFunction(Vbig)
     qpts = Q.get_points()
     qwts = Q.get_weights()
 
@@ -78,76 +67,64 @@ def getFormDiscGalerkin(F, L, Q, t, dt, u0, bcs=None, nullspace=None):
     perm = [*edofs[0][0], *edofs[1][0], *edofs[0][1]]
     basis_vals = basis_vals[perm]
     basis_dvals = basis_dvals[perm]
+    basis_vals_w = np.multiply(basis_vals, qwts)
 
     # mass matrix later for BC
-    mmat = np.multiply(basis_vals, qwts) @ basis_vals.T
-
+    mmat = basis_vals_w @ basis_vals.T
     # L2 projector
-    proj = Constant(np.linalg.solve(mmat, np.multiply(basis_vals, qwts)))
+    proj = vecconst(np.linalg.solve(mmat, basis_vals_w))
 
-    u_np = np.reshape(UU, (num_stages, *u0.ufl_shape))
-    v_np = np.reshape(VV, (num_stages, *u0.ufl_shape))
+    trial_vals = vecconst(basis_vals)
+    trial_dvals = vecconst(basis_dvals)
+    test_vals_w = vecconst(basis_vals_w)
+    qpts = vecconst(qpts.reshape((-1,)))
 
     split_form = extract_terms(F)
-    dtless = strip_dt_form(split_form.time)
+    F_dtless = strip_dt_form(split_form.time)
+    F_remainder = split_form.remainder
 
-    Fnew = zero()
+    # set up the pieces we need to work with to do our substitutions
+    v_np = np.reshape(test, (num_stages, *u0.ufl_shape))
+    u_np = np.reshape(stages, (num_stages, *u0.ufl_shape))
+    vsub = test_vals_w.T @ v_np
+    usub = trial_vals.T @ u_np
+    dtu0sub = trial_dvals.T @ u_np
 
-    basis_vals = vecconst(basis_vals)
-    basis_dvals = vecconst(basis_dvals)
-
-    qpts = vecconst(qpts.reshape((-1,)))
-    qwts = vecconst(qwts)
-
-    # Terms with time derivatives
-    for i in range(num_stages):
-        repl = {v: v_np[i]}
-        F_i = component_replace(dtless, repl)
-
-        # now loop over quadrature points
-        for q in range(len(qpts)):
-            repl = {t: t + dt * qpts[q],
-                    u0: (1/dt) * (u_np @ basis_dvals[:, q])}
-
-            Fnew += dt * qwts[q] * basis_vals[i, q] * component_replace(F_i, repl)
-
-    # jump terms
+    # Jump terms
     repl = {u0: u_np[0] - u0,
             v: v_np[0]}
+    Fnew = component_replace(F_dtless, repl)
 
-    Fnew += component_replace(dtless, repl)
+    # Terms with time derivatives
+    for q in range(len(qpts)):
+        repl = {t: t + qpts[q] * dt,
+                v: vsub[q] * dt,
+                u0: dtu0sub[q] / dt}
+        Fnew += component_replace(F_dtless, repl)
 
-    # handle the rest of the terms
-    for i in range(num_stages):
-        repl = {v: v_np[i]}
-        F_i = component_replace(split_form.remainder, repl)
-
-        # now loop over quadrature points
-        for q in range(len(qpts)):
-            repl = {t: t + dt * qpts[q],
-                    u0: u_np @ basis_vals[:, q]}
-
-            Fnew += dt * qwts[q] * basis_vals[i, q] * component_replace(F_i, repl)
+    # Handle the rest of the terms
+    for q in range(len(qpts)):
+        repl = {t: t + qpts[q] * dt,
+                v: vsub[q] * dt,
+                u0: usub[q]}
+        Fnew += component_replace(F_remainder, repl)
 
     # Oh, honey, is it the boundary conditions?
     if bcs is None:
         bcs = []
     bcsnew = []
     for bc in bcs:
-        bcarg = as_ufl(bc._original_arg)
-        bcblah_at_qp = np.zeros((len(qpts),), dtype="O")
-        for q in range(len(qpts)):
-            tcur = t + qpts[q] * dt
-            bcblah_at_qp[q] = replace(bcarg, {t: tcur})
-        bc_func_for_stages = dot(proj, as_vector(bcblah_at_qp))
+        g0 = as_ufl(bc._original_arg)
+        Vg_np = np.array([replace(g0, {t: t + c*dt}) for c in qpts])
+        g_np = proj @ Vg_np
         for i in range(num_stages):
             Vbigi = stage2spaces4bc(bc, V, Vbig, i)
-            bcsnew.append(bc.reconstruct(V=Vbigi, g=bc_func_for_stages[i]))
+            bcsnew.append(bc.reconstruct(V=Vbigi, g=g_np[i]))
 
-    return Fnew, UU, bcsnew, getNullspace(V, Vbig, num_stages, nullspace)
+    return Fnew, bcsnew
 
 
-class DiscontinuousGalerkinTimeStepper:
+class DiscontinuousGalerkinTimeStepper(StageCoupledTimeStepper):
     """Front-end class for advancing a time-dependent PDE via a Discontinuous Galerkin
     in time method
 
@@ -191,10 +168,6 @@ class DiscontinuousGalerkinTimeStepper:
                  quadrature=None,
                  solver_parameters=None, appctx=None, nullspace=None):
         assert order >= 0
-        self.u0 = u0
-        self.orig_bcs = bcs
-        self.t = t
-        self.dt = dt
         self.order = order
         self.basis_type = basis_type
 
@@ -219,52 +192,17 @@ class DiscontinuousGalerkinTimeStepper:
         self.quadrature = quadrature
         assert np.size(quadrature.get_points()) >= order+1
 
-        self.num_steps = 0
-        self.num_nonlinear_iterations = 0
-        self.num_linear_iterations = 0
+        num_stages = order+1
+        super().__init__(F, t, dt, u0, num_stages, bcs=bcs,
+                         solver_parameters=solver_parameters,
+                         appctx=appctx, nullspace=nullspace)
 
-        bigF, UU, bigBCs, bigNSP = \
-            getFormDiscGalerkin(F, self.el,
-                                quadrature, t, dt, u0, bcs, nullspace)
+    def get_form_and_bcs(self, stages):
+        return getFormDiscGalerkin(self.F, self.el,
+                                   self.quadrature, self.t, self.dt, self.u0, stages,
+                                   self.orig_bcs)
 
-        self.UU = UU
-        self.bigBCs = bigBCs
-        problem = NLVP(bigF, UU, bigBCs)
-        appctx_irksome = {"F": F,
-                          "t": t,
-                          "dt": dt,
-                          "u0": u0,
-                          "bcs": bcs,
-                          "nullspace": nullspace}
-        if appctx is None:
-            appctx = appctx_irksome
-        else:
-            appctx = {**appctx, **appctx_irksome}
-
-        push_parent(u0.function_space().dm, UU.function_space().dm)
-        self.solver = NLVS(problem,
-                           appctx=appctx,
-                           solver_parameters=solver_parameters,
-                           nullspace=bigNSP)
-        pop_parent(u0.function_space().dm, UU.function_space().dm)
-
-    def advance(self):
-        """Advances the system from time `t` to time `t + dt`.
-        Note: overwrites the value `u0`."""
-        push_parent(self.u0.function_space().dm, self.UU.function_space().dm)
-        self.solver.solve()
-        pop_parent(self.u0.function_space().dm, self.UU.function_space().dm)
-
-        self.num_steps += 1
-        self.num_nonlinear_iterations += self.solver.snes.getIterationNumber()
-        self.num_linear_iterations += self.solver.snes.getLinearSolveIterations()
-
-        u0 = self.u0
-        u0bits = u0.subfunctions
-        UUs = self.UU.subfunctions
-
-        for i, u0bit in enumerate(u0bits):
-            u0bit.assign(UUs[self.num_fields*(self.order)+i])
-
-    def solver_stats(self):
-        return (self.num_steps, self.num_nonlinear_iterations, self.num_linear_iterations)
+    def _update(self):
+        # only correct if we are nodal w.r.t. right end.
+        for i, u0bit in enumerate(self.u0.subfunctions):
+            u0bit.assign(self.stages.subfunctions[self.num_fields*(self.order)+i])
