+++ irksome/galerkin_stepper.py
@@ -1,19 +1,17 @@
-from functools import reduce
 from FIAT import (Bernstein, DiscontinuousElement, DiscontinuousLagrange,
                   IntegratedLegendre, Lagrange, Legendre,
                   make_quadrature, ufc_simplex)
-from operator import mul
 from ufl import zero
 from ufl.constantvalue import as_ufl
+from .base_time_stepper import StageCoupledTimeStepper
 from .bcs import bc2space, stage2spaces4bc
 from .deriv import TimeDerivative
-from .tools import getNullspace, ConstantOrZero, component_replace, replace
+from .tools import component_replace, replace, vecconst
 import numpy as np
-from firedrake import as_vector, dot, Constant, TestFunction, Function, NonlinearVariationalProblem as NLVP, NonlinearVariationalSolver as NLVS
-from firedrake.dmhooks import pop_parent, push_parent
+from firedrake import TestFunction
 
 
-def getFormGalerkin(F, L_trial, L_test, Q, t, dt, u0, bcs=None, nullspace=None):
+def getFormGalerkin(F, L_trial, L_test, Q, t, dt, u0, stages, bcs=None):
 
     """Given a time-dependent variational form, trial and test spaces, and
     a quadrature rule, produce UFL for the Galerkin-in-Time method.
@@ -29,23 +27,16 @@ def getFormGalerkin(F, L_trial, L_test, Q, t, dt, u0, bcs=None, nullspace=None):
          The user may adjust this value between time steps.
     :arg u0: a :class:`Function` referring to the state of
          the PDE system at time `t`
+    :arg stages: a :class:`Function` representing the stages to be solved for.
     :arg bcs: optionally, a :class:`DirichletBC` object (or iterable thereof)
          containing (possibly time-dependent) boundary conditions imposed
          on the system.
-    :arg nullspace: A list of tuples of the form (index, VSB) where
-         index is an index into the function space associated with `u`
-         and VSB is a :class: `firedrake.VectorSpaceBasis` instance to
-         be passed to a `firedrake.MixedVectorSpaceBasis` over the
-         larger space associated with the Runge-Kutta method
 
     On output, we return a tuple consisting of four parts:
 
        - Fnew, the :class:`Form` corresponding to the Galerkin-in-Time discretized problem
-       - UU, the :class:`Function` representing the stages to be solved for
        - `bcnew`, a list of :class:`firedrake.DirichletBC` objects to be posed
          on the Galerkin-in-time solution,
-       - 'nspnew', the :class:`firedrake.MixedVectorSpaceBasis` object
-         that represents the nullspace of the coupled system
     """
     assert L_test.get_reference_element() == Q.ref_el
     assert L_trial.get_reference_element() == Q.ref_el
@@ -57,60 +48,48 @@ def getFormGalerkin(F, L_trial, L_test, Q, t, dt, u0, bcs=None, nullspace=None):
     assert V == u0.function_space()
 
     num_stages = L_test.space_dimension()
-
-    Vbig = reduce(mul, (V for _ in range(num_stages)))
-    VV = TestFunction(Vbig)
-    UU = Function(Vbig)  # u0 + this are coefficients of the Galerkin polynomial
-
-    u_np = np.reshape(UU, (num_stages, *u0.ufl_shape))
-    v_np = np.reshape(VV, (num_stages, *u0.ufl_shape))
-
+    Vbig = stages.function_space()
+    test = TestFunction(Vbig)
     qpts = Q.get_points()
     qwts = Q.get_weights()
-
-    tabulate_trials = L_trial.tabulate(1, qpts)
-    trial_vals = tabulate_trials[(0,)]
-    trial_dvals = tabulate_trials[(1,)]
-    test_vals = L_test.tabulate(0, qpts)[(0,)]
 
     # sort dofs geometrically by entity location
     edofs = L_trial.entity_dofs()
     trial_perm = [*edofs[0][0], *edofs[1][0], *edofs[0][1]]
-    trial_vals = trial_vals[trial_perm]
-    trial_dvals = trial_dvals[trial_perm]
+    tabulate_trials = L_trial.tabulate(1, qpts)
+    trial_vals = tabulate_trials[(0,)][trial_perm]
+    trial_dvals = tabulate_trials[(1,)][trial_perm]
+    test_vals = L_test.tabulate(0, qpts)[(0,)]
+    test_vals_w = np.multiply(test_vals, qwts)
 
     # mass-ish matrix later for BC
-    mmat = np.multiply(test_vals, qwts) @ trial_vals[1:].T
-
+    mmat = test_vals_w @ trial_vals[1:].T
     # L2 projector
-    proj = Constant(np.linalg.solve(mmat, np.multiply(test_vals, qwts)))
+    proj = vecconst(np.linalg.solve(mmat, test_vals_w))
 
-    Fnew = zero()
-
-    vecconst = np.vectorize(ConstantOrZero)
     trial_vals = vecconst(trial_vals)
     trial_dvals = vecconst(trial_dvals)
-    test_vals = vecconst(test_vals)
+    test_vals_w = vecconst(test_vals_w)
     qpts = vecconst(qpts.reshape((-1,)))
-    qwts = vecconst(qwts)
 
-    for i in range(num_stages):
-        repl = {v: v_np[i]}
-        F_i = component_replace(F, repl)
+    # set up the pieces we need to work with to do our substitutions
+    v_np = np.reshape(test, (num_stages, *u0.ufl_shape))
+    w_np = np.reshape(stages, (num_stages, *u0.ufl_shape))
+    u_np = np.concatenate((np.reshape(u0, (1, *u0.ufl_shape)), w_np))
+    vsub = test_vals_w.T @ v_np
+    usub = trial_vals.T @ u_np
+    dtu0sub = trial_dvals.T @ u_np
 
-        # now loop over quadrature points
-        for q in range(len(qpts)):
-            tosub = u0 * trial_vals[0, q]
-            tosub += sum(u_np[j] * trial_vals[1+j, q] for j in range(num_stages))
+    dtu0 = TimeDerivative(u0)
 
-            d_tosub = u0 * trial_dvals[0, q]
-            d_tosub += sum(u_np[j] * trial_dvals[1+j, q] for j in range(num_stages))
-
-            repl = {t: t + dt * qpts[q],
-                    u0: tosub,
-                    TimeDerivative(u0): d_tosub / dt}
-
-            Fnew += dt * qwts[q] * test_vals[i, q] * component_replace(F_i, repl)
+    # now loop over quadrature points
+    Fnew = zero()
+    for q in range(len(qpts)):
+        repl = {t: t + qpts[q] * dt,
+                v: vsub[q] * dt,
+                u0: usub[q],
+                dtu0: dtu0sub[q] / dt}
+        Fnew += component_replace(F, repl)
 
     # Oh, honey, is it the boundary conditions?
     if bcs is None:
@@ -118,20 +97,18 @@ def getFormGalerkin(F, L_trial, L_test, Q, t, dt, u0, bcs=None, nullspace=None):
     bcsnew = []
     for bc in bcs:
         u0_sub = bc2space(bc, u0)
-        bcarg = as_ufl(bc._original_arg)
-        bcblah_at_qp = np.zeros((len(qpts),), dtype="O")
-        for q in range(len(qpts)):
-            tcur = t + qpts[q] * dt
-            bcblah_at_qp[q] = replace(bcarg, {t: tcur}) - u0_sub * trial_vals[0, q]
-        bc_func_for_stages = dot(proj, as_vector(bcblah_at_qp))
+        g0 = as_ufl(bc._original_arg)
+        Vg_np = np.array([replace(g0, {t: t + c * dt}) for c in qpts])
+        Vg_np -= u0_sub * trial_vals[0]
+        g_np = proj @ Vg_np
         for i in range(num_stages):
             Vbigi = stage2spaces4bc(bc, V, Vbig, i)
-            bcsnew.append(bc.reconstruct(V=Vbigi, g=bc_func_for_stages[i]))
+            bcsnew.append(bc.reconstruct(V=Vbigi, g=g_np[i]))
 
-    return Fnew, UU, bcsnew, getNullspace(V, Vbig, num_stages, nullspace)
+    return Fnew, bcsnew
 
 
-class GalerkinTimeStepper:
+class GalerkinTimeStepper(StageCoupledTimeStepper):
     """Front-end class for advancing a time-dependent PDE via a Galerkin
     in time method
 
@@ -207,52 +184,14 @@ class GalerkinTimeStepper:
         self.quadrature = quadrature
         assert np.size(quadrature.get_points()) >= order
 
-        self.num_steps = 0
-        self.num_nonlinear_iterations = 0
-        self.num_linear_iterations = 0
+        super().__init__(F, t, dt, u0, order, bcs=bcs,
+                         solver_parameters=solver_parameters,
+                         appctx=appctx, nullspace=nullspace)
 
-        bigF, UU, bigBCs, bigNSP = \
-            getFormGalerkin(F, self.trial_el, self.test_el,
-                            quadrature, t, dt, u0, bcs, nullspace)
+    def get_form_and_bcs(self, stages):
+        return getFormGalerkin(self.F, self.trial_el, self.test_el,
+                               self.quadrature, self.t, self.dt, self.u0, stages, self.orig_bcs)
 
-        self.UU = UU
-        self.bigBCs = bigBCs
-        problem = NLVP(bigF, UU, bigBCs)
-        appctx_irksome = {"F": F,
-                          "t": t,
-                          "dt": dt,
-                          "u0": u0,
-                          "bcs": bcs,
-                          "nullspace": nullspace}
-        if appctx is None:
-            appctx = appctx_irksome
-        else:
-            appctx = {**appctx, **appctx_irksome}
-
-        push_parent(u0.function_space().dm, UU.function_space().dm)
-        self.solver = NLVS(problem,
-                           appctx=appctx,
-                           solver_parameters=solver_parameters,
-                           nullspace=bigNSP)
-        pop_parent(u0.function_space().dm, UU.function_space().dm)
-
-    def advance(self):
-        """Advances the system from time `t` to time `t + dt`.
-        Note: overwrites the value `u0`."""
-        push_parent(self.u0.function_space().dm, self.UU.function_space().dm)
-        self.solver.solve()
-        pop_parent(self.u0.function_space().dm, self.UU.function_space().dm)
-
-        self.num_steps += 1
-        self.num_nonlinear_iterations += self.solver.snes.getIterationNumber()
-        self.num_linear_iterations += self.solver.snes.getLinearSolveIterations()
-
-        u0 = self.u0
-        u0bits = u0.subfunctions
-        UUs = self.UU.subfunctions
-
-        for i, u0bit in enumerate(u0bits):
-            u0bit.assign(UUs[self.num_fields*(self.order-1)+i])
-
-    def solver_stats(self):
-        return (self.num_steps, self.num_nonlinear_iterations, self.num_linear_iterations)
+    def _update(self):
+        for i, u0bit in enumerate(self.u0.subfunctions):
+            u0bit.assign(self.stages.subfunctions[self.num_fields*(self.order-1)+i])
