+++ irksome/imex.py
@@ -8,8 +8,8 @@ from ufl import zero
 
 from .ButcherTableaux import RadauIIA
 from .deriv import TimeDerivative
-from .stage import getFormStage
-from .tools import AI, ConstantOrZero, IA, MeshConstant, replace, component_replace
+from .stage_value import getFormStage
+from .tools import AI, ConstantOrZero, IA, MeshConstant, replace, component_replace, getNullspace, get_stage_space
 from .bcs import bc2space
 
 
@@ -181,9 +181,17 @@ class RadauIIAIMEXMethod:
 
         # Since this assumes stiff accuracy, we drop
         # the update information on the floor.
-        Fbig, _, UU, bigBCs, nsp = getFormStage(
-            F, butcher_tableau, u0, t, dt, bcs,
-            splitting=splitting, nullspace=nullspace)
+        V = u0.function_space()
+        Vbig = get_stage_space(V, self.num_stages)
+        UU = Function(Vbig)
+
+        Fbig, bigBCs = getFormStage(
+            F, butcher_tableau, t, dt, u0, UU, bcs,
+            splitting=splitting)
+
+        nsp = getNullspace(u0.function_space(),
+                           UU.function_space(),
+                           self.num_stages, nullspace)
 
         self.UU = UU
         self.UU_old = UU_old = Function(UU.function_space())
